Add permissions to the company

+++ backend/company/views.py
@@ -3,9 +3,12 @@ from rest_framework.response import Response
 from rest_framework import status
 from .models import Company, Catalog, Product
 from .serializers import CompanySerializer, CatalogSerializer, ProductSerializer
-
+from .permissions import *
 
 class CompanyListCreateView(APIView):
+    
+    permission_classes = (AdminLevelPermission, )
+    
     def get(self, request):
         companies = Company.objects.all()
         serializer = CompanySerializer(companies, many=True)
@@ -26,13 +29,34 @@ class CompanyListCreateView(APIView):
 
 
 class CompanyRetrieveUpdateDeleteView(APIView):
+    
+    permission_classes = (AdminLevelPermission, )
+    
     def get(self, request, id):
+        
+        user = request.user
+        
+        
         company = Company.objects.get(pk=id)
+        
+        if user.id != company.company_owner.id and (not user.is_superadmin):
+            return Response({
+                "error": "You do not have permission for this company"
+            }, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = CompanySerializer(company)
         return Response(serializer.data)
 
     def patch(self, request, id):
+        user = request.user
+        
         company = Company.objects.get(pk=id)
+        
+        if user.id != company.company_owner.id and (not user.is_superadmin):
+            return Response({
+                "error": "You do not have permission for this company"
+            }, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = CompanySerializer(company, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -40,12 +64,24 @@ class CompanyRetrieveUpdateDeleteView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
+        
+        user = request.user
+        
         company = Company.objects.get(pk=id)
+        
+        if user.id != company.company_owner.id and (not user.is_superadmin):
+            return Response({
+                "error": "You do not have permission for this company"
+            }, status=status.HTTP_403_FORBIDDEN)
+            
         company.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class ProductListCreateView(APIView):
+    
+    permission_classes = (EmployeeLevelPermission, )
+    
     def get(self, request, company_id):
         products = Product.objects.filter(company=company_id)
         serializer = ProductSerializer(products, many=True)
